+++ pydidas_plugins/proc_plugins/pyfai_2d_integration.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2023, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -16,20 +18,20 @@
 """
 Module with the PyFAI2dIntegration Plugin which allows to integrate diffraction
 patterns into a 2D radial/azimuthal map.
-
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
-__status__ = "Development"
+__status__ = "Production"
 __all__ = ["PyFAI2dIntegration"]
+
+
+import numpy as np
 
 from pydidas.core import Dataset
 from pydidas.plugins import pyFAIintegrationBase
-
-import numpy as np
 
 
 class PyFAI2dIntegration(pyFAIintegrationBase):
@@ -101,6 +103,7 @@ class PyFAI2dIntegration(pyFAIintegrationBase):
         kwargs : dict
             Any calling kwargs, appended by any changes in the function.
         """
+        self.check_and_set_custom_mask(**kwargs)
         _newdata = self._ai.integrate2d(
             data, self.get_param_value("rad_npoint"), **self._ai_params
         )